src/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
boto3